feat: add pocket remote app / show versions for apps - add RemoteService usage to computer apps (pocket/turtle to be done) - allows remote reboot/update - add remote configurability for subway-switch max distance

+++ app/pocket/bone-meal.lua
@@ -1,8 +1,17 @@
-package.path = package.path .. ";/?.lua"
+if package then
+    package.path = package.path .. ";/?.lua"
+end
+
+local version = require "version"
+
+if not arg then
+    return version
+end
+
 local Rpc = require "lib.common.rpc"
 local BoneMealService = require "lib.features.bone-meal-service"
 
-print("[bone-meal v1.3.0-dev] booting...")
+print(string.format("[bone-meal %s] booting...", version()))
 local on = arg[1] == "on"
 local off = arg[1] == "off"
 local reboot = arg[1] == "reboot"
